Move QgsOrientedBoundingBox out to own file
