+++ python/core/core_auto.sip
@@ -345,6 +345,7 @@
 %Include auto_generated/geometry/qgsmultipoint.sip
 %Include auto_generated/geometry/qgsmultipolygon.sip
 %Include auto_generated/geometry/qgsmultisurface.sip
+%Include auto_generated/geometry/qgsorientedboundingbox.sip
 %Include auto_generated/geometry/qgspoint.sip
 %Include auto_generated/geometry/qgspolygon.sip
 %Include auto_generated/geometry/qgsquadrilateral.sip
