+++ src/core/tiledmesh/qgscesiumutils.h
@@ -29,87 +29,7 @@ using namespace nlohmann;
 #endif
 
 class QgsSphere;
-
-/**
- * \brief Represents a oriented (rotated) bounding box in 3 dimensions.
- *
- * \ingroup core
- *
- * \warning Non-stable API, exposed to Python for unit testing only.
- *
- * \since QGIS 3.34
- */
-class CORE_EXPORT QgsOrientedBoundingBox
-{
-  public:
-#ifndef SIP_RUN
-
-    /**
-     * Creates an oriented bounding box from a Cesium json object.
-     */
-    static QgsOrientedBoundingBox fromJson( const json &json );
-#endif
-
-    /**
-     * Constructor for a null bounding box.
-     */
-    QgsOrientedBoundingBox();
-
-    /**
-     * Constructor for a oriented bounding box, with a specified center and half axes matrix.
-     */
-    QgsOrientedBoundingBox( const QList<double> &center, QList< double > &halfAxes );
-
-    /**
-     * Returns TRUE if the box is a null bounding box.
-     */
-    bool isNull() const;
-
-    /**
-     * Returns the center x-coordinate.
-     *
-     * \see centerY()
-     * \see centerZ()
-     */
-    double centerX() const { return mCenter[0]; }
-
-    /**
-     * Returns the center y-coordinate.
-
-     * \see centerX()
-     * \see centerZ()
-     */
-    double centerY() const { return mCenter[1]; }
-
-    /**
-     * Returns the center z-coordinate.
-     *
-     * \see centerX()
-     * \see centerY()
-     */
-    double centerZ() const { return mCenter[2]; }
-
-    /**
-     * Returns the half axes matrix;
-     */
-    const double *halfAxes() const SIP_SKIP { return mHalfAxes; }
-
-    /**
-     * Returns the half axes matrix;
-     */
-    QList< double > halfAxesList() const SIP_PYNAME( halfAxes );
-
-    /**
-     * Returns the overall bounding box of the object.
-     */
-    QgsBox3d extent() const;
-
-  private:
-
-    double mCenter[ 3 ] { std::numeric_limits< double >::quiet_NaN(), std::numeric_limits< double >::quiet_NaN(), std::numeric_limits< double >::quiet_NaN() };
-    double mHalfAxes[9] { 1, 0, 0, 0, 1, 0, 0, 0, 1 };
-
-};
+class QgsOrientedBoundingBox;
 
 /**
  * \brief Contains utilities for working with Cesium data.
