+++ tests/src/python/CMakeLists.txt
@@ -252,6 +252,7 @@ ADD_PYTHON_TEST(PyQgsOGRProviderGpkg test_provider_ogr_gpkg.py)
 ADD_PYTHON_TEST(PyQgsOGRProviderSqlite test_provider_ogr_sqlite.py)
 ADD_PYTHON_TEST(PyQgsOpacityWidget test_qgsopacitywidget.py)
 ADD_PYTHON_TEST(PyQgsOptional test_qgsoptional.py)
+ADD_PYTHON_TEST(PyQgsOrientedBoundingBox test_qgsorientedboundingbox.py)
 ADD_PYTHON_TEST(PyQgsOwsConnection test_qgsowsconnection.py)
 ADD_PYTHON_TEST(PyQgsPalLabelingBase test_qgspallabeling_base.py)
 ADD_PYTHON_TEST(PyQgsPalLabelingCanvas test_qgspallabeling_canvas.py)
